Separate OpenAI-compatible provider tests into dedicated test suite
